rename: AI 퀴즈 생성 관련 컴포넌트 파일 위치 변경
